IDLのクラス名を修正 refs #3
